+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 ### New Features
 
+* Add 1Password SecureStorage backend (`SecureStore: 1password`) with service account and desktop
+  app authentication support #1369
 * Add aws-sso [exec|eval] --overwrite-env flag to allow overriding AWS_ env variables #1095
 * Add support for /healthcheck endpoint for ECS server #1356
 * Add `--default <profile>` flag to `ecs server` and `ecs docker start` to automatically load